+++ src/App.js
@@ -9,16 +9,8 @@ import AuthForm from "./Components/AuthForm";
 import Container from "react-bootstrap/Container";
 import Navbar from "react-bootstrap/Navbar";
 import { Routes, Route, Link, Navigate } from "react-router-dom";
-import {
-  onChildAdded,
-  ref as databaseRef,
-  onChildRemoved,
-} from "firebase/database";
-import { database } from "./firebase";
 
 export const UserContext = React.createContext({});
-// Save the Firebase post folder name as a constant to avoid bugs due to misspelling
-const POST_KEY = "posts";
 
 function RequireAuth({ children, redirectTo, loggedInUser }) {
   console.log(loggedInUser);
@@ -39,36 +31,6 @@ function App() {
       }
     });
   }, [loggedInUser]);
-
-  // Initialise empty posts array in state to keep local state in sync with Firebase
-  // When Firebase changes, update local state, which will update local UI
-  const [posts, setPosts] = useState([]);
-
-  useEffect(() => {
-    const postRef = databaseRef(database, POST_KEY);
-    // onChildAdded will return data for every child at the reference and every subsequent new child
-    onChildAdded(postRef, (data) => {
-      console.log(data);
-      // Add the subsequent child to local component state, initialising a new array to trigger re-render
-      setPosts((state) =>
-        // Store posts key so we can use it as a key in our card items when rendering posts
-        [...state, { key: data.key, val: data.val() }]
-      );
-    });
-  }, []);
-
-  useEffect(() => {
-    const postRef = databaseRef(database, POST_KEY);
-    // Delete the deleted post & image from state so it is no longer rendered when mounted
-    onChildRemoved(postRef, (removedOldData) => {
-      console.log("data onChildRemoved", removedOldData);
-      const postsCopy = [...posts];
-      const newPosts = postsCopy.filter(
-        (post) => post.key !== removedOldData.key
-      );
-      setPosts(newPosts);
-    });
-  }, [posts]);
 
   const handleLikeCount = () => {
     setLikeCount((state) =>
@@ -130,7 +92,7 @@ function App() {
                     <Composer email={loggedInUser.email} />
                   </RequireAuth>
 
-                  <NewsFeed onclick={handleLikeButton} posts={posts} />
+                  <NewsFeed onclick={handleLikeButton} />
                 </>
               }
             />
@@ -142,7 +104,7 @@ function App() {
                   <RequireAuth loggedInUser={loggedInUser}>
                     <Composer email={loggedInUser.email} />
                   </RequireAuth>
-                  <NewsFeed onClick={handleLikeButton} posts={posts} />
+                  <NewsFeed onClick={handleLikeButton} />
                 </>
               }
             />
