+++ src/Components/NewsFeed.js
@@ -2,18 +2,54 @@ import Card from "react-bootstrap/Card";
 import { Button } from "react-bootstrap";
 import { ref as storageRef, deleteObject } from "firebase/storage";
 import { database, storage } from "../firebase";
-import { ref as databaseRef, remove } from "firebase/database";
+import {
+  ref as databaseRef,
+  remove,
+  onChildAdded,
+  onChildRemoved,
+} from "firebase/database";
+import { useState, useEffect } from "react";
 
 const POST_KEY = "posts";
 
-function NewsFeed({ posts, handleDelete }) {
-  // Convert posts in props to JSX elements to render
+function NewsFeed() {
+  // Initialise empty posts array in state to keep local state in sync with Firebase
+  // When Firebase changes, update local state, which will update local UI
+  const [posts, setPosts] = useState([]);
+
+  useEffect(() => {
+    const postRef = databaseRef(database, POST_KEY);
+    // onChildAdded will return data for every child at the reference and every subsequent new child
+    onChildAdded(postRef, (data) => {
+      console.log(data);
+      // Add the subsequent child to local component state, initialising a new array to trigger re-render
+      setPosts((state) =>
+        // Store posts key so we can use it as a key in our card items when rendering posts
+        [...state, { key: data.key, val: data.val() }]
+      );
+    });
+  }, []);
+
+  useEffect(() => {
+    const postRef = databaseRef(database, POST_KEY);
+    // Delete the deleted post & image from state so it is no longer rendered
+    onChildRemoved(postRef, (removedOldData) => {
+      console.log("data onChildRemoved", removedOldData);
+      const postsCopy = [...posts];
+      const newPosts = postsCopy.filter(
+        (post) => post.key !== removedOldData.key
+      );
+      setPosts(newPosts);
+    });
+  }, [posts]);
+
+  // Convert posts in state to JSX elements to render
   console.log(posts);
   let postItems = posts.map((post) => (
     <Card key={post.key}>
       {post.val.url ? (
         <>
-          <Card.Img variant="top" src={post.val.url} className="Card-img" />{" "}
+          <Card.Img variant="top" src={post.val.url} className="Card-img" />
           <Button variant="white" onClick={onclick}>
             ❤ {post.val.likeCount}
           </Button>
