+++ src/firebase.js
@@ -9,8 +9,7 @@ const firebaseConfig = {
   apiKey: process.env.REACT_APP_API_KEY,
   authDomain: process.env.REACT_APP_AUTH_DOMAIN,
   // The value of `databaseURL` depends on the location of the database
-  databaseURL:
-    "https://rocketgram-ad72d-default-rtdb.asia-southeast1.firebasedatabase.app/",
+  databaseURL: process.env.REACT_APP_DATABASE_URL,
   projectId: process.env.REACT_APP_PROJECT_ID,
   storageBucket: process.env.REACT_APP_STORAGE_BUCKET,
   messagingSenderId: process.env.REACT_APP_MESSAGING_SENDER_ID,
